✨ feat(内核调度, syscall 协程, 时钟中断, per-hart运行时上下文): 重构等待与协程执行链路并添加per-hart运行时上下文
 * 协程抽象与生命周期管理
    - 重构 util::cotask，补充 void 协程特化，统一返回值与无返回值协程的行为
    - 引入句柄所有权与 final_suspend 清理逻辑，避免 detached 协程悬挂或重复释放
    - 调整 awaiter 语义，使协程在恢复、等待和销毁之间的切换更明确
 * hart 级运行时上下文
    - 为 env 增加 HartContext，按 hart 保存当前线程、进程、运行队列、trap 上下文和 timer
    - 将 tmm / trap_context 等运行时状态从全局逻辑下沉到每个 hart 私有上下文
    - 补充 reset_runtime 等接口，便于 hart 初始化、切换和重置运行状态
 * 时钟与中断控制链路
    - 新增 clock 抽象和 CSRTimeClockSource，统一提供 tick / 时间换算能力
    - 新增 ClintTimer，并把 CLINT 中断、timebase-frequency、DeviceModel 注册流程串联起来
    - 在 FDT 解析中补充 CPU 频率、CLINT 目标 hart 和控制器注册的校验与日志
 * syscall、trap 与等待流程
    - 重写 syscall 调度入口，增加 active_context、同步分发和可挂起 syscall 的恢复逻辑
    - 重新整理 trap.S / entry.S / hart.S / isr.cpp 的入口与上下文保存恢复路径
    - 调整 wait/scheduler/task 相关实现，让等待线程、syscall 协程和重新入队之间的状态衔接更完整
 * 配套测试与接口更新
    - 补充 coroutine 相关测试，用于覆盖新协程语义和恢复行为
    - 同步更新 cap、uaccess、task、scheduler、endpoint、memory 等接口签名与调用链
    - 清理旧的上下文实现，替换为新的架构层实现与 include 配置

+++ kernel/arch/riscv64/setup.cpp
@@ -19,9 +19,10 @@
 #include <sus/logger.h>
 #include <sus/units.h>
 
+#include <cstddef>
 #include <cstring>
 
-int hart_id;
+size_t hart_id;
 void *dtb_ptr;
 
 namespace
